commit before pairprogramming
