+++ frontend/src/communications/apiInstance.js
@@ -1,5 +1,6 @@
 import axios from 'axios'
 
+
 const api = axios.create({
   baseURL: `/api/`,
 })
