+++ frontend/src/communications/userApi.js
@@ -1,12 +1,11 @@
 import api from './apiInstance'
-import './mockForUserApi'
 
 export const getUsers = async () => {
   try {
     const response = await api.get('/users')
     return response.data.users
   } catch (error) {
-    //throw new Error('Failed to get users.')
+    throw new Error('Failed to get users.')
   }
 }
 
@@ -23,5 +22,18 @@ export const sendApproval = async (userId) => {
     return await api.put(`/users/${userId}/approval`)
   } catch (error) {
     throw new Error('Registration approval miscarried.')
+  }
+}
+
+export const getProducts = async () => {
+  try {
+    const response = await api.get('/products')
+    
+      console.log(response.data.products);
+        
+    return response.data.products
+    
+  } catch (error) {
+    throw new Error('Failed to get products.')
   }
 }
